Modified background classes for #69.
