+++ src/guis/krow/guis/GUIHelper.java
@@ -33,7 +33,8 @@ import javafx.scene.text.FontWeight;
 import javafx.scene.text.Text;
 import javafx.scene.text.TextAlignment;
 import javafx.util.Duration;
-import krow.guis.BackgroundBuilder.ShapeBackgroundManager;
+import krow.backgrounds.BackgroundBuilder;
+import krow.backgrounds.ShapeBackgroundManager;
 import kröw.core.Kröw;
 import kröw.core.managers.WindowManager;
 import zeale.guis.Home;
