+++ src/krow/kröw/core/managers/WindowManager.java
@@ -14,12 +14,11 @@ import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Label;
-import javafx.scene.layout.Pane;
 import javafx.scene.paint.Color;
 import javafx.stage.Popup;
 import javafx.stage.Stage;
 import javafx.util.Duration;
-import krow.guis.BackgroundBuilder.BackgroundManager;
+import krow.backgrounds.BackgroundManager;
 import kröw.core.Kröw;
 import kröw.events.Event;
 import kröw.events.EventHandler;
